Bug Fixing

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -16,7 +16,7 @@ class CreateUsersTable extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id()->unique(); 
             $table->integer('role');
-            $table->integer('nomor_induk');
+            $table->bigInteger('nomor_induk');
             $table->string('name');
             $table->string('email');
             $table->timestamp('email_verified_at')->nullable();
